Try to get the last known location from every LocationManager based LocationProvider
The GpsStatusListener-Code was removed, because the documentation says that event with id = 4
will be called periodically. That would mean that every second we get a new location and with every
new location the signals would be triggered when a thunderstorm is close.

Apart from that, not only we can get the last known location for GPS, but also for the other LocationManager
based LocationProviders